+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- [**BC**] Use net6.0
+- [**BC**] Remove deprecated `configure` in Consumer configuration
 
 ## 18.0.0 - 2021-10-29
 - [**BC**] Add MetaData types